requirements.txt
python-dotenv
pymongo
sentence-transformers
faiss-cpu
gradio
openai
numpy
pydantic
